Week2/DailyProject/InvoiceExtraction/requirements.txt
streamlit
pytesseract
pillow
pdf2image
groq
python-dotenv
opencv-python